flask_app/requirements.txt
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.1
flask-jwt-extended[asymmetric_crypto]==4.6.0
pem==23.1.0
bcrypt==4.2.0
pytz==2024.2
pytest==7.4.2
coverage==7.6.7

black==23.12.1
cryptography==42.0.4
Flask==3.0.1
pre-commit==3.6.0
PyJWT==2.8.0
tox==4.12.1
waitress==3.0.1
twilio==9.4.1
